backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.14
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt
dnspython
APScheduler==3.10.4
python-dotenv
email-validator
# Async high-performance dependencies
aiodns>=3.1.1
aiosmtplib>=3.0.1
cachetools>=5.3.0
